+++ NovaChat/app/src/main/java/com/novachat/ui/navigation/NavRoutes.kt
@@ -70,3 +70,6 @@ object TrustedSendersRoute
 
 @Serializable
 object InboxSpamScanRoute
+
+@Serializable
+object MessagingSettingsRoute